Separate DIBuilder API from Numba IR API
Marking a subprogram, variable, or location in the DIBuilder API
requires passing in a `numba.core.ir.Loc` object. This makes sense when
all debuginfo generation comes from lowering, but creates weird
dependencies when generating code that doesn't relate to a specific
piece of IR - for example, the kernel wrapper generated in the CUDA
target. To mark locations in the kernel wrapper, one must import the
`Loc` class from `numba.core.ir`, and construct instances of it to pass
in, just to mark locations in generated code (or do something more odd
with a "pretend" `Loc` object).

Since the `Loc` instances are only used for the line number, this commit
changes the interface of the relevant `DIBuilder` methods to accept a
line number instead of a `Loc`, breaking the spurious coupling.