CUDA: Explicitly specify objmode + looplifting for jit functions in cuda.random
Using the `@jit` decorator alone results in warnings being emitted when
using the `cuda.random` functions with the simulator enabled, because
Fake CUDA Arrays are passed to jitted functions. Since these can't be
typed or unboxed, a fallback to object mode with looplifting is needed.

To avoid the warning, we explicitly specify that object mode with
looplifting should be used when the simulator is enabled.

Fixes #4809.