CUDA: Fix ffs
The `ffs` intrinsic produced results that were off-by-1 compared to
`__ffs()` in CUDA C (and in general compared with `ffs()`
implementations in C libraries). This commit fixes the implementation by
making use of the libdevice implementations `__nv_ffs` and `__nv_ffsll`
for 32- and 64-bit operands respectively. Since these always return a
32-bit integer between 0 and 32 (0 for no bit set), the typing of the
function is modified to always return a `uint32`.

Tests are updated accordingly.

The documentation is updated to clarify the index of the least
significant bit, and the result of calling `ffs(0)`. Documentation for
other bit manipulation instructions is tidied up and made consistent
with the documentation for `ffs`.

Fixes #6919.