Remove const record optimization tests
These tests check for very specific patterns in the generated PTX that
are unstable across toolkit versions, platforms, and other minor
changes. Since these frequently need adjustment to match the generated
PTX, and have never revealed an underlying issue, is seems more sensible
to remove them instead of continually fixing them up in the presence of
variations.

Note that the latest issue was detected in the PR #6840 buildfarm run,
with failures in test_const_record_optimization on Windows for CUDA 9.2
and 10.0. These failures arose because a change to compile the kernel
with an explicit signature instead of specializing the kernel for given
arguments changed the types the kernel was being compiled for on Windows
- the default int on Windows is different to the default int on Linux.
So this test wasn't even testing the same thing on Windows as on Linux,
and this went unnoticed for a very long time!