CUDA: Skip multi-GPU copy test with peer access disabled
When peer access is disabled (e.g. with one GPU in WDDM mode and one in
TCC mode), a device-to-device copy between the two fails, causing
`TestMultiGPUContext.test_with_context` to fail. We fix this issue by
moving the test of device-to-device copy into its own test method, which
is skipped when peer access is disabled.

Fixes #7253.