Use correct capitalisation in struct.calcsize
uintc was being set based on the size of struct.calcsize('i') (int), and
similarly size_t was based on the size of struct.calcsize('n')
(ssize_t). Fix them to I and N. Should have no functional impact since
unsigned versions of types should always be the same size, but
technically correct is the best kind of correct.