CUDA: Only retain exception message when Driver.__init__ catches CudaSupportError
This addresses a further cause of a memory leak similar to the previous
commit. Additional tests are added with CUDA disabled to force an error
during `Driver.__init__()`.

A very small edit is made to `initialization_error_test()`, because
`cuda_error()` should be available any time, not just when catching an
exception from initializing CUDA. This does not affect the test, but
does more closely mirror any potential use cases.