CUDA: Type device functions as dispatchers
This brings the typing of device functions closer in line with the way
it works for the CPU target. The root of the change is that now
`get_overload()` returns a unique identifier for the overload instead of
the compile result; on the CPU target, the entry point is returned as
the overload, but there is no entry point on CUDA, so the `id()` of the
compile result is used instead, as it is the only simple and unique
property of CUDA compile results.

This forces other structural changes, mostly in that the
`DeviceDispatcher` (lazily-compiled device functions) and
`DeviceFunction` (eagerly-compiled device functions) are now typed as
dispatchers. Other code changes (addition of new functions and
properties) are needed because the core of Numba expects their presence
in objects typed as dispatchers.

A CUDA Dispatcher type is needed, because the lowering of dispatchers
in general attempts to lower them to a unique address for first-class
function support. As there is no first-class function support in CUDA,
the lowering of its dispatchers needs to produce a dummy value instead.

One test is updated due to a change in the error message that it scans.