Add py_func to FakeCUDAKernel (fix record dtype test)
The lack of a py_func in the FakeCUDAKernel is a slightly surprising
omission, although it must have been missed for a long time as many CUDA
tests have to be skipped with the simulator.

Since it brings the API into line with the hardware target and fixes a
record dtype test without requiring any complex / invasive changes, we
add it here.