Handle PTDS with CUDA Python at the CUDA Python layer
Per-thread default streams are handled by CUDA Python directly and are
not the responsibility of Numba when using CUDA Python. Therefore we
don't need to look up the PTDS APIs with CUDA Python, and shouldn't
tests for them either.

Users requesting PTDS in conjunction with CUDA Python will be directed
to set the appropriate CUDa Python environment variable instead.