Permit a new array model to have a super set of array model fields.
In the case where a user defines a new array model that extends
from or duck types to Numba's default array model, it should be
valid for the new array model to have additional fields and still
be accepted by Numba's array related implementations.

This patch fixes a check to permit the reuse of the somewhat
fundamental ```numba.np.arrayobj.populate_array``` function.