Fix `isinstance` warning check test.
The `isinstance` warning is raised from an `@overload`, as a
result its appearance is stateful with respect to the use of
`isinstance` in a JIT function. This patch pushes the test into
a subprocess to guarantee the state, i.e. has not been used in
the subprocess so the warning will always appear.